chore(client):rename folder Roles to ResourceRoles
